Fixed add stock
