+++ src/main/java/app/carport/Persistence/MaterialMapper.java
@@ -9,6 +9,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Provides database operations for managing materials within the application.
@@ -161,6 +162,50 @@ public class MaterialMapper {
 
         } catch (SQLException e) {
             throw new DatabaseException("Error. Couldn't get material from the given materialID.", e.getMessage());
+        }
+    }
+
+    /**
+     * Updates the stock quantities of materials in the database by subtracting the specified quantities.
+     * This method iterates over a map of materials and their quantities to be removed, updating the database accordingly.
+     *
+     * @param materialList   A map containing Material objects as keys and integers representing the quantity of each material to be removed.
+     * @param connectionPool The connection pool from which to obtain database connections.
+     * @return true if the update operation completes successfully for all materials.
+     * @throws DatabaseException If a SQLException occurs during database update operations.
+     */
+    public static boolean removeMaterialStockOnOrder(Map<Material, Integer> materialList, ConnectionPool connectionPool) throws DatabaseException {
+        String sql = "UPDATE material SET \"quantityInStock\" = \"quantityInStock\" - ? WHERE \"name\" = ?";
+        try (Connection connection = connectionPool.getConnection(); PreparedStatement ps = connection.prepareStatement(sql)) {
+            for (Map.Entry<Material, Integer> entry : materialList.entrySet()) {
+                ps.setInt(1, entry.getValue()); // quantityToRemove
+                ps.setString(2, entry.getKey().getName()); // materialName
+                ps.executeUpdate();
+            }
+            return true;
+        } catch (SQLException e) {
+            throw new DatabaseException("Error. Couldn't remove material stock from the database.", e.getMessage());
+        }
+    }
+
+    /**
+     * Updates the stock quantity of a specific material in the database.
+     *
+     * @param materialID     The ID of the material to update.
+     * @param quantityToAdd  The amount to add to the current stock quantity.
+     * @param connectionPool The connection pool to use for obtaining a database connection.
+     * @return {@code true} if the stock was updated successfully; {@code false} otherwise.
+     * @throws DatabaseException If a database error occurs.
+     */
+    public static boolean addMaterialStock(int materialID, int quantityToAdd, ConnectionPool connectionPool) throws DatabaseException {
+        String sql = "UPDATE material SET \"quantityInStock\" = \"quantityInStock\" + ? WHERE \"materialID\" = ?";
+        try (Connection connection = connectionPool.getConnection(); PreparedStatement ps = connection.prepareStatement(sql)) {
+            ps.setInt(1, quantityToAdd);
+            ps.setInt(2, materialID);
+            int affectedRows = ps.executeUpdate();
+            return affectedRows > 0;
+        } catch (SQLException e) {
+            throw new DatabaseException("Error. Couldn't update material stock in the database.", e.getMessage());
         }
     }
 }
